Added missing new line in csv and EnviroDIY header
Also made the loggingInterval a consistent uint16_t (it was a mix of uint8_t and float)

+++ src/LoggerDreamHost.cpp
@@ -15,7 +15,7 @@
 // ============================================================================
 
 // Constructor
-LoggerDreamHost::LoggerDreamHost(const char *loggerID, uint8_t loggingIntervalMinutes,
+LoggerDreamHost::LoggerDreamHost(const char *loggerID, uint16_t loggingIntervalMinutes,
                                  int8_t SDCardPin, int8_t mcuWakePin,
                                  VariableArray *inputArray)
   : LoggerEnviroDIY(loggerID, loggingIntervalMinutes, SDCardPin, mcuWakePin, inputArray)
